Reuse AggregateState across ZStream and StreamEffect
